+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.1
 
 - Adds scroll controllers for picked items & showed items.
+- Adds scroll physics for picked items & showed items.
 
 ## 1.0.0